requirements.txt
fastapi==0.124.4
uvicorn[standard]==0.38.0
pydantic==2.12.5
python-dotenv==1.2.1
python-multipart==0.0.12
groq==0.13.0
google-generativeai==0.8.3
pillow>=11.1.0
httpx>=0.24.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
fastapi-mail>=1.4.1
authlib>=1.3.0
itsdangerous>=2.1.2

